% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_hankel.R
\name{build_hankel}
\alias{build_hankel}
\title{Build a Hankel matrix from a time series.}
\usage{
build_hankel(x, stackmax)
}
\arguments{
\item{x}{A vector of measurements over time.}

\item{stackmax}{An integer; the number of shift-stacked rows.}
}
\value{
A Hankel matrix of \code{x}.
}
\description{
Create a Hankel matrix from a vector of measurements over time.
}
\examples{
\dontrun{
build_hankel(x = xdat, stackmax = 15)
}
}
