% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_derivative.R
\name{compute_derivative}
\alias{compute_derivative}
\title{Compute numeric derivatives of a time series.}
\usage{
compute_derivative(x, dt, r = min(dim(as.matrix(x))),
                          devMethod = c("FOCD", "GLLA", "FINITE"),
                          gllaEmbed = NA, gllaTau = NA, gllaOrder = NA)
}
\arguments{
\item{x}{A vector or matrix of measurements over time.}

\item{dt}{Numeric; the change in time between successive measurements.}

\item{r}{An integer; the number of time series in \code{x} used to calculate \code{dXdt}.}

\item{devMethod}{A character string. One of either \code{"FOCD"} for fourth order central difference, \code{"GLLA"} for generalized local linear approximation,
or \code{"FINITE"} for simple finite difference.}

\item{gllaEmbed}{An integer; t embedding dimension used for \code{devMethod = "GLLA"}.}

\item{gllaTau}{An integer; the time delay used for \code{devMethod = "GLLA"}.}

\item{gllaOrder}{An integer; the embedding dimension used for \code{devMethod = "GLLA"}.}
}
\value{
\itemize{
\item If \code{devMethod = "FOCD"} - returns a vector or matrix of first order derivatives the first \code{r} columns of \code{x} with respect to time.
\item If \code{devMethod = "GLLA"} - returns a matrix or list of matrices of derivatives of the first \code{r} columns of \code{x} with respect to time.
Derivatives returned by \code{devMethod = "GLLA"} are up to order \code{"gllaOrder"} and include the "\eqn{0^{th}}" derivative.
\item If \code{devMethod = "FINITE"} - returns a vector or matrix of first order derivatives the first \code{r} columns of \code{x} with respect to time.
}
}
\description{
Estimates the derivative of a time series using numeric methods.
}
\examples{
data(ECG_measurements)

xdat <- ECG_measurements[,"channel1"]
dt <- ECG_measurements[2,"time"] - ECG_measurements[1,"time"]

compute_derivative(x = xdat, dt = dt, devMethod = "FOCD")

\dontrun{
compute_derivative(x = xdat, dt = dt, devMethod = "GLLA", gllaEmbed = 5, gllaTau = 1, gllaOrder = 1)
}
}
\references{
Boker, S. M., Deboeck, P. R., Edler, C., & Keel, P. K. (2010). Generalized local linear approximation of derivatives
 from time series.
 In Chow S, Ferrer E, and Hsieh F, editors, Statistical methods for modeling human dynamics: An interdisciplinary dialogue.
}
