% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmd.R
\name{dmd}
\alias{dmd}
\title{Perform Dynamic Mode Decomposition (DMD) on a matrix of snapshots of a system.}
\usage{
dmd(x, y = NULL, r, dt)
}
\arguments{
\item{x}{A matrix of snapshots of a system.}

\item{y}{A matrix of snapshot paired to \code{x}.}

\item{r}{An integer; the specific number of singular vectors to include.}

\item{dt}{Numeric; the time-lag between two subsequent time series measurements.}
}
\value{
An object of class 'dmd' with the following components: \itemize{
\item{\code{dmdResult} - }{TBD}
\item{\code{timeDynamics} - }{TBD}
\item{\code{dmdModes} - }{TBD}
\item{\code{dmdAmplitudes} - }{TBD}
\item{\code{dtEigen} - }{TBD}
\item{\code{ctEigen} - }{TBD}}
}
\description{
Dynamic mode composition (DMD) is a method for approximating the eigenvalue and eigenvectors of
 the Koopman operator of a system. That is, DMD yields the eigendecomposition of the linear map of a system
 from time \eqn{t} to time \eqn{t+1}.
}
\examples{
library(pracma)

# Generate data
x <- seq(-5, 5, length.out = 128)
t <- seq(0, 4*pi, length.out = 256)

grids <- meshgrid(x, t)

# First periodic function
f1xt <- t(sech(grids$X + 3)*exp(2.3i*grids$Y))

# Second periodic function
f2xt <- t(2*sech(grids$X)*tanh(grids$X)*exp(2.8i*grids$Y))

# Observed values
fxt <- f1xt + f2xt

dmd(fxt, r = 2, dt = 1)
}
\references{
Kutz, J. N., Brunton, S. L., Brunton, B. W., & Proctor, J. L. (2016).
Dynamic mode decomposition: data-driven modeling of complex systems.
Society for Industrial and Applied Mathematics.
}
