% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECG_measurements-data.R
\docType{data}
\name{ECG_measurements}
\alias{ECG_measurements}
\title{Two channels of ECG measurements obtained from the QT database}
\format{A data frame with 45000 rows and 3 variables:
\describe{
  \item{time}{Time of measurment in seconds}
  \item{channel1}{ECG reading of first channel}
  \item{channel2}{ECG reading of second channel}
  }}
\source{
\href{https://physionet.org/content/qtdb/1.0.0/}{QT database}
}
\usage{
data(ECG_measurements)
}
\description{
Data from record sel102. 45,000 measurements of 2 ECG channels over 180 seconds.
}
\examples{
\donttest{
data(ECG_measurements)
}
}
\references{
Laguna, P., Mark, R. G., Goldberg, A., & Moody, G. B.
 (1997, September). A database for evaluation of algorithms for measurement
  of QT and other waveform intervals in the ECG. In Computers in cardiology
   1997 (pp. 673-676). IEEE.
}
\keyword{datasets}
