% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internet_Trolls-data.R
\docType{data}
\name{Internet_Trolls}
\alias{Internet_Trolls}
\title{Topic scores of right-wing Russian troll Twitter activity}
\format{
A data frame with 1453 rows and 16 variables:
\describe{
  \item{Date}{TO-DO}
  \item{Time_of_day}{TO-DO}
  \item{Type}{TO-DO}
  \item{Topic1}{TO-DO}
  \item{Topic2}{TO-DO}
  \item{Topic3}{TO-DO}
  \item{Topic4}{TO-DO}
  \item{Topic5}{TO-DO}
  \item{Topic6}{TO-DO}
  \item{Topic7}{TO-DO}
  \item{Topic8}{TO-DO}
  \item{Topic9}{TO-DO}
  \item{Topic10}{TO-DO}
  \item{Topic11}{TO-DO}
  \item{author}{TO-DO}
  \item{Time_stamp}{TO-DO}
  }
}
\source{
\href{https://github.com/fivethirtyeight/russian-troll-tweets/}{www.fivethirtyeight.com's github}
}
\usage{
data(Internet_Trolls)
}
\description{
1453 scores on 11 Topics.
}
\examples{
\donttest{
data(Internet_Trolls)
}
}
\keyword{datasets}
