% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmd.R
\name{mldivide}
\alias{mldivide}
\title{Dynamic Mode Decomposition (DMD)}
\usage{
mldivide(A, B, pinv = TRUE)
}
\arguments{
\item{x}{A matrix of snapshots}

\item{y}{A matrix of snapshot paired to \code{x}}

\item{r}{An integer; specific number of singular vectors to include.}

\item{dt}{A numeric value indicating the time-lag between two subsequent time series measurements.}
}
\value{

}
\description{
DMD Description
}
\examples{
\dontrun{
sparsify_dynamics(Theta, dXdt, lambda, n)
sparsify_dynamics(Theta, dXdt, 0.1, 10)
sparsify_dynamics(pool_data(yIn, 15, 5, TRUE), dXdt, 0, 15)
}
}
\references{
Brunton, S. L., Proctor, J. L., & Kutz, J. N. (2016). Discovering
governing equations from data by sparse identification of nonlinear dynamical
systems. Proceedings of the National Academy of Sciences, 113(15), 3932-3937.
}
