% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_SVHT_coef.R
\name{optimal_SVHT_coef}
\alias{optimal_SVHT_coef}
\title{Matrix Sparcification for SINDy Algorithm}
\usage{
optimal_SVHT_coef(beta, sigma_known = FALSE)
}
\arguments{
\item{beta}{A number.}

\item{sigma_known}{A number.}
}
\value{
A matrix of sparse coefficients
}
\description{
Coefficient determining optimal location of Hard Threshold for Matrix
 Denoising by Singular Values Hard Thresholding when noise level is known or
 unknown.  Recreation of matlab code by Matan Gavish and David Donoho.
}
\examples{
add(1, 1)
add(10, 1)
}
