% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_sparsify_dynamics.R
\name{par_sparsify_dynamics}
\alias{par_sparsify_dynamics}
\title{Matrix Sparsification for SINDy Algorithm}
\usage{
par_sparsify_dynamics(Theta, dXdt, lambda, loops)
}
\arguments{
\item{Theta}{A matrix of candidate functions.}

\item{dXdt}{A matrix of first order derivatives of the variables of interest with respect to time.}

\item{lambda}{A numeric value; sparsification threshold.}

\item{loops}{ADD DESCRIPTION}
}
\value{
A matrix of sparse coefficients.
}
\description{
Sparsification function based on sequential thresholded least-squares
as shown in the SINDy algorithm in "Discovering governing equations from data:
Sparse identification of nonlinear dynamical systems" (Brunton, Proctor, & Kutz, 2016).
}
\examples{
\dontrun{
sparsify_dynamics(Theta, dXdt, lambda, n)
sparsify_dynamics(Theta, dXdt, 0.1, 10)
sparsify_dynamics(pool_data(yIn, 15, 5, TRUE), dXdt, 0, 15)
}
}
\references{
Brunton, S. L., Proctor, J. L., & Kutz, J. N. (2016). Discovering
governing equations from data by sparse identification of nonlinear dynamical
systems. Proceedings of the National Academy of Sciences, 113(15), 3932-3937.
}
