% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_align.R
\name{svd_align}
\alias{svd_align}
\title{Align singular vectors with majority of data vectors}
\usage{
svd_align(x, align = TRUE)
}
\arguments{
\item{x}{A matrix of measurements.}
}
\value{
\code{svd_flip} Aligned svd
}
\description{
The svd algorithm is specified up to sign (+/-). This algorithm aligns the sign of the singular vectors with the data.
}
\examples{
# Generate data
x <- cbind(runif(10, -0.5, 2), runif(10, -0.5, 2))


svd_flip <- svd_align(x)

# Extract components of svd and plot
U <- svd_flip$u
Sigma <- svd_flip$d
VT <- t(svd_flip$v)

# Plot data
plot(0, 0, xlim=c(-2,2), ylim=c(-2,2), type="n")

for (i in 1:dim(x)[1]){
  arrows(0, 0, x[i,1], x[i,2], lwd=2)
}

arrows(0, 0, VT[1, 1], VT[1, 2], col='red')

}
\references{
Bro, R., Acar, E., & Kolda, T. G. (2008). Resolving the sign ambiguity in the singular value decomposition. Journal of Chemometrics: A Journal of the Chemometrics Society, 22(2), 135-140.
}
