% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_forcing.R
\name{active_forcing}
\alias{active_forcing}
\title{Determine when forcing is active for an object of class "havok"}
\usage{
active_forcing(x, minVal = stats::sd(x$x[, x$r]))
}
\arguments{
\item{x}{An object of class "havok"}

\item{minVal}{A cutoff value for determining when forcing is active. Defualts
to one standard deviation of the forcing term.}
}
\value{
A matrix of sparse coefficients.
}
\description{
This function uses a minimum value to determin when forcing
is active in a fitted "havok" object.
}
\examples{
\donttest{
hav <- havok(xdat = xdat, dt = dt)
active_forcing(x = hav)
}
}
