% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_derivative.R
\name{compute_derivative}
\alias{compute_derivative}
\title{Compute derivatives of a time series.}
\usage{
compute_derivative(x, dt, r = min(dim(as.matrix(x))),
                          devMethod = c("FOCD", "GLLA"),
                          gllaEmbed = NA, gllaTau = NA, gllaOrder = NA)
}
\arguments{
\item{x}{A vector or matrix of measurements over time.}

\item{dt}{A number representing the change in time between successive measurements.}

\item{r}{An integer representing the number of time series in \code{x} used to calculate \code{dXdt}.}

\item{devMethod}{A character string. One of either \code{"FOCD"} for fourth order central difference or \code{"GLLA"} for generalized local linear approximation.}

\item{gllaEmbed}{An integer representing the embedding dimension used for \code{devMethod = "GLLA"}.}

\item{gllaTau}{An integer representing the time delay used for \code{devMethod = "GLLA"}.}

\item{gllaOrder}{An integer representing the embedding dimension used for \code{devMethod = "GLLA"}.}
}
\value{
\itemize{
\item If \code{devMethod = "FOCD"} - returns a vector or matrix of first order derivatives the first \code{r} columns of \code{x} with respect to time.
\item If \code{devMethod = "GLLA"} - returns a matrix or list of matrices of derivatives of the first \code{r} columns of \code{x} with respect to time.
Derivatives returned by \code{devMethod = "GLLA"} are up to order \code{"gllaOrder"} and include the "\eqn{0^{th}}" derivative.
}
}
\description{
Estimates the derivative of a time series using fourth order central difference or
generalized local linear approximation.
}
\examples{
data(ECG_measurements)

xdat <- ECG_measurements[,"channel1"]
dt <- ECG_measurements[2,"time"] - ECG_measurements[1,"time"]

compute_derivative(x = xdat, dt = dt, devMethod = "FOCD")

\dontrun{
compute_derivative(x = xdat, dt = dt, devMethod = "GLLA", gllaEmbed = 5, gllaTau = 1, gllaOrder = 1)
}
}
