% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_data.R
\name{pool_data}
\alias{pool_data}
\title{Data Pooling for SINDy Algorithm}
\usage{
pool_data(yIn, nVars, polyOrder, useSine)
}
\arguments{
\item{yIn}{A number.}

\item{nVars}{A number.}

\item{polyOrder}{A number.}

\item{useSine}{A number.}
}
\value{
A matrix of candidate functions
}
\description{
Pooling function for sparsifyIng dynamics as shown in
the SINDy algorithm as shown in "Discovering
governing equations from data: Sparse identification of nonlinear dynamical
systems" (Brunton, Proctor, & Kutz, 2015).
}
\examples{
add(1, 1)
add(10, 1)
}
