% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sindy.R
\name{sindy}
\alias{sindy}
\title{Sparse Identification of Nonlinear Dynamics (SINDy)}
\usage{
sindy(
  x,
  dt,
  lambda,
  polyOrder = 5,
  useSine = FALSE,
  normalize = FALSE,
  nVars = ncol(as.matrix(x))
)
}
\arguments{
\item{x}{A vector or matrix of measurments over time.}

\item{dt}{A numeric value indicating the time-lag between two subsequent time series measures.}

\item{lambda}{A numeric value; sparsification threshold.}
}
\value{
A matrix of sparse coefficients.
}
\description{
Sparsification function based on sequential thresholded least-squares
as shown in the SINDy algorithm in "Discovering governing equations from data:
Sparse identification of nonlinear dynamical systems" (Brunton, Proctor, & Kutz, 2016).
}
\examples{
\donttest{
sparsify_dynamics(Theta, dXdt, lambda, n)
sparsify_dynamics(Theta, dXdt, 0.1, 10)
sparsify_dynamics(pool_data(yIn, 15, 5, TRUE), dXdt, 0, 15)
}
}
\references{
Brunton, S. L., Proctor, J. L., & Kutz, J. N. (2016). Discovering
governing equations from data by sparse identification of nonlinear dynamical
systems. Proceedings of the National Academy of Sciences, 113(15), 3932-3937.
}
